Add dry run

+++ delete.py
@@ -13,7 +13,7 @@ def _filter(name):
     )
 
 
-def _delete(images):
+def _delete(images, dry_run):
     for i in images:
         get_snapshot_id = lambda: [
             d["Ebs"]["SnapshotId"] for d in i.block_device_mappings if "Ebs" in d
@@ -22,22 +22,28 @@ def _delete(images):
         name = i.name
         snapshot_id = get_snapshot_id()
 
-        i.deregister()
-        print(name + " deregistered")
+        print(name + " deregistering")
 
-        ec2.Snapshot(snapshot_id).delete()
-        print(snapshot_id + " deleted")
+        if not dry_run:
+            i.deregister()
+            print(name + " deregistered")
+
+        print(snapshot_id + " deleting")
+
+        if not dry_run:
+            ec2.Snapshot(snapshot_id).delete()
+            print(snapshot_id + " deleted")
 
 
-def delete_tail(name):
+def delete_tail(name, dry_run):
     images = _filter(name)
 
     _, *tail = sorted(images, key=lambda i: i.creation_date, reverse=True)
 
-    _delete(tail)
+    _delete(tail, dry_run)
 
 
-def delete_days(name, days):
+def delete_days(name, days, dry_run):
     images = _filter(name)
 
     today = datetime.now(timezone.utc).replace(
@@ -46,4 +52,4 @@ def delete_days(name, days):
 
     test = lambda i: (today - parser.isoparse(i.creation_date)).days >= days
 
-    _delete([i for i in images if test(i)])
+    _delete([i for i in images if test(i)], dry_run)
